backend/requirements.txt
Django>=5,<6
djangorestframework>=3.14,<4
psycopg2-binary>=2.9,<3
django-cors-headers>=4.6,<5
factory_boy>=3.3.3,<3.4
django-extensions>=3.2.3,<3.3
drf-yasg>=1.21.8,<2
django-phonenumber-field[phonenumbers]==8.0.0
